requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
contentful==2.4.0
elastic-transport==8.17.1
elasticsearch==9.0.0
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
